requirements.txt
mysql-connector-python==8.0.27
pandas==1.3.5
pytz==2021.3
